Revert "Align Orders queries with existing use case view models"
