+++ src/EChamado/EChamado.Shared/Domain/Events/BrighterEventMapper.cs
@@ -1,13 +1,13 @@
-using EChamado.Shared.Domain;
-using Paramore.Brighter;
+﻿using Paramore.Brighter;
 
-namespace EChamado.Server.Infrastructure.Events;
+namespace EChamado.Shared.Domain.Events;
 
 public sealed class BrighterEventMapper : IBrighterEventMapper
 {
     private readonly Dictionary<Type, Func<IDomainEvent, IRequest>> _map = new();
 
-    public BrighterEventMapper Register<TDomainEvent>(Func<TDomainEvent, IRequest> factory)
+    public BrighterEventMapper Register<TDomainEvent>(
+        Func<TDomainEvent, IRequest> factory)
         where TDomainEvent : IDomainEvent
     {
         _map[typeof(TDomainEvent)] = e => factory((TDomainEvent)e);
@@ -17,9 +17,7 @@ public sealed class BrighterEventMapper : IBrighterEventMapper
     public IRequest? Map(IDomainEvent domainEvent)
     {
         if (_map.TryGetValue(domainEvent.GetType(), out var factory))
-        {
             return factory(domainEvent);
-        }
 
         return null;
     }
