+++ src/EChamado/EChamado.Shared/Domain/Events/IDomainEventDispatcher.cs
@@ -1,0 +1,6 @@
+﻿namespace EChamado.Shared.Domain.Events;
+
+public interface IDomainEventDispatcher
+{
+    Task DispatchAsync(IEnumerable<IDomainEvent> domainEvents, CancellationToken ct);
+}
