+++ src/EChamado/EChamado.Shared/EChamado.Shared.csproj
@@ -9,6 +9,7 @@
   <ItemGroup>
     <PackageReference Include="FluentValidation" Version="12.1.0" />
     <PackageReference Include="LinqKit.Microsoft.EntityFrameworkCore" Version="9.0.8" />
+    <PackageReference Include="Paramore.Brighter" Version="10.0.3" />
   </ItemGroup>
