--- src/EChamado/Server/EChamado.Server.Application/Orders/Queries/GetOrderByIdQuery.cs
@@ -1,14 +1,0 @@
-using EChamado.Server.Application.UseCases.Orders.ViewModels;
-using Paramore.Darker;
-
-namespace EChamado.Server.Application.Orders.Queries;
-
-public sealed class GetOrderByIdQuery : IQuery<OrderViewModel?>
-{
-    public Guid OrderId { get; }
-
-    public GetOrderByIdQuery(Guid orderId)
-    {
-        OrderId = orderId;
-    }
-}
