--- src/EChamado/Server/EChamado.Server.Infrastructure/Events/IDomainEventDispatcher.cs
@@ -1,8 +1,0 @@
-using EChamado.Shared.Domain;
-
-namespace EChamado.Server.Infrastructure.Events;
-
-public interface IDomainEventDispatcher
-{
-    Task DispatchAsync(IEnumerable<IDomainEvent> domainEvents, CancellationToken cancellationToken);
-}
